+++ src/pydvl/value/shapley/common.py
@@ -46,12 +46,9 @@ def compute_shapley_values(
       Shapley. Computation is **not parallelized**. Implemented in
       :func:`~pydvl.value.shapley.naive.permutation_exact_shapley`.
     - ``permutation_montecarlo``: uses the approximate Monte Carlo
-      implementation of permutation data Shapley. Implemented in
+      implementation of permutation data Shapley. Accepts a
+      :class:`TruncationPolicy` to stop computing marginals. Implemented in
       :func:`~pydvl.value.shapley.montecarlo.permutation_montecarlo_shapley`.
-    - ``truncated_montecarlo``: default option, same as ``permutation_montecarlo``
-      but stops the computation whenever a certain accuracy is reached.
-      Implemented in
-      :func:`~pydvl.value.shapley.montecarlo.truncated_montecarlo_shapley`.
     - ``owen_sampling``: Uses the Owen continuous extension of the utility
       function to the unit cube. Implemented in
       :func:`~pydvl.value.shapley.montecarlo.owen_sampling_shapley`. This
@@ -98,7 +95,11 @@ def compute_shapley_values(
     if mode not in list(ShapleyMode):
         raise ValueError(f"Invalid value encountered in {mode=}")
 
-    if mode == ShapleyMode.TruncatedMontecarlo:
+    if mode in (
+        ShapleyMode.PermutationMontecarlo,
+        ShapleyMode.ApproShapley,
+        ShapleyMode.TruncatedMontecarlo,
+    ):
         truncation = kwargs.pop("truncation", NoTruncation())
         return truncated_montecarlo_shapley(  # type: ignore
             u=u, done=done, n_jobs=n_jobs, truncation=truncation, **kwargs
@@ -106,16 +107,6 @@ def compute_shapley_values(
     elif mode == ShapleyMode.CombinatorialMontecarlo:
         return combinatorial_montecarlo_shapley(
             u, done=done, n_jobs=n_jobs, progress=progress
-        )
-    elif mode in (ShapleyMode.PermutationMontecarlo, ShapleyMode.ApproShapley):
-        truncation = kwargs.pop("truncation", NoTruncation())
-        return permutation_montecarlo_shapley(
-            u,
-            done=done,
-            n_jobs=n_jobs,
-            progress=progress,
-            truncation=truncation,
-            **kwargs,
         )
     elif mode == ShapleyMode.CombinatorialExact:
         return combinatorial_exact_shapley(u, n_jobs=n_jobs, progress=progress)
